autonomousの微調整
